requirements.txt
# LangChain and LangGraph
langchain-openai>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.40

# Azure and Authentication
azure-identity>=1.15.0
requests>=2.31.0

# Data Fetching
yfinance>=0.2.0
pandas>=2.0.0

# Environment Management
python-dotenv>=1.0.0
